Large refactor
- Rename load_jpeg to load_jpeg_chw
- Rename some more input files by order where applicable
- Minor reformat
- Implement selector for xcorr/conv in own CL convolution
- Change sepconv-cl from conv to xcorr
- Improve OpenCL kernel doc
- Implement Coeff / CoeffFloat for f64
- Make softmax consuming
- Remove some extra asserts
- Generalize verify()
- Abbreviate ndarray::Array::permuted_axes as reorder()

+++ src/cl/mtx_mul.cl
@@ -7,14 +7,17 @@
 #include "common.h"
 #include "hyperparams.h"
 
-
+// Does matrix multiplication: ´weights´x´input´=´output´
+// input/B: (channels, height, width) [CHW]
+// output/c_mul: () [const]
+// weights/A: (channels, height, width, const) [CHWN]
 __kernel void mtx_mul(
     __global CL_PRIM_N* restrict input,
     __global CL_PRIM* restrict output,
     __global CL_PRIM_N* restrict weights) {
 
     // Divide by number of primitives in vector
-    const int DATA_SIZE = PATCH3SQ * FM_COUNT / VECN;
+    const size_t DATA_SIZE = PATCH3SQ * FM_COUNT / VECN;
 
     size_t gid = get_global_id(0);
